+++ packages/hydrogen/src/components/LocalizationProvider/LocalizationProvider.server.tsx
@@ -1,9 +1,18 @@
-import React, {ReactNode} from 'react';
+import React, {ReactNode, ElementType} from 'react';
 import LocalizationClientProvider from './LocalizationClientProvider.client';
 import {useShopQuery} from '../../hooks/useShopQuery';
 import {LocalizationQuery} from './LocalizationQuery';
 import {Localization} from '../../graphql/graphql-constants';
 import {CacheDays} from '../../framework/CachingStrategy';
+import {PreloadOptions} from '../../types';
+import {Props} from '../types';
+
+export interface LocalizationProviderProps {
+  /** Any ReactNode elements. */
+  children: ReactNode;
+  /** `true` to preload the localization query for the url or `'*'` to preload the localization query for all urls */
+  preload: PreloadOptions;
+}
 
 /**
  * The `LocalizationProvider` component automatically queries the Storefront API's
@@ -14,17 +23,20 @@ import {CacheDays} from '../../framework/CachingStrategy';
  * The `isoCode` of the `country` can be used in the Storefront API's
  * `@inContext` directive as the `country` value.
  */
-export function LocalizationProvider({children}: {children: ReactNode}) {
+export function LocalizationProvider<TTag extends ElementType>(
+  props: Props<TTag> & LocalizationProviderProps
+) {
   const {
     data: {localization},
   } = useShopQuery<LocalizationQuery>({
     query: Localization,
     cache: CacheDays(),
+    preload: props.preload,
   });
 
   return (
     <LocalizationClientProvider localization={localization}>
-      {children}
+      {props.children}
     </LocalizationClientProvider>
   );
 }